backend/port_scanner/requirements.txt
pymongo
neo4j-driver
python-dotenv